go: Delegate responsibility to free C-allocated memory to caller
This fixes an issue where struct members were passed to external C
functions after which the struct was no longer referenced.
As described in the docs of runtime.SetFinalizer, this would allow
the Go garbage collector to run every finalizer associated with the
struct (as it will no longer be used). Unfortunately, as we bound
a finalizer to free every C-allocated struct member, those
finalizers were run, freeing the struct members too early.

This patch currently requires the caller to manually call .Free()
on the following structs once they are no longer required:
- Context
- ImageHandle
- DecodingOptions
- Image

Such structs are currently returned by the following functions:
- NewContext
- NewDecodingOptions
- NewImage
- Context.GetPrimaryImageHandle
- Context.GetImageHandle
- ImageHandle.GetDepthImageHandle
- ImageHandle.GetThumbnail
- ImageHandle.DecodeImage
- Image.ScaleImage