go: Fix C-allocated struct members being freed too early
This fixes an issue where struct members were passed to external C
functions after which the struct was no longer referenced.
As described in the docs of runtime.SetFinalizer, this would allow
the Go garbage collector to run every finalizer associated with the
struct (as it will no longer be used). Unfortunately, as we bound
a finalizer to free every C-allocated struct member, those
finalizers were run, freeing the struct members too early.