Add temporary installation directory in macOS build instructions
My motivation was not depend on the repository maintainers to get new binaries with bug fixes. In my case the compilation instructions for macOS work but the executables in the 'examples' directory are wrappers. Because I needed the binaries I added the 'prefix'-flag in the macOS build instructions. Hopefully now it should be transparent where to find fresh compilated binaries. This could decrease the dependence of users on the repository maintainers.