change project name to heif
This change is done so that it will produce:

Unix: libheif.so
Windows: heif.lib & heif.dll

The include directory is hardcoded to libheif so that consumers of the library does not need to
change anything to find the header files. It will only change for windows consumers, which will need
to update the library name from libheif to heif.

From my point of view, it is legit to make this change because anyways, any CMake consumer of libheif
should now use the CMake Config file to find the dependency, and that config file is in charge of
indicating the library name and include directories. Note that it is also the default behaviour in
CMake to have libXXX.so for UNIX and XXX.lib on Windows. Experiments trying to modify the default
behavior end up in additional code for the maintained project, and additional difficulties for the
consumers of the library.