Add support for the AOM key=value API
This allows callers to pass custom options to the AOM encoder by adding
an 'aom:' prefix to the option name.
Th custom options will be applied after libheif configures the default
options because they can override the values that were set.

The x265 encoder already has a similar feature with the 'x265:' prefix.

Resolves #596