heif_encoder_aom: change defaults of speed, min-q
Change the default of the "speed" parameter from 5 to 6.

Change the default of the "min-q" parameter from 1 to 0 (the default of
cfg.rc_min_quantizer in libaom). Also change the range of the "min-q"
and "alpha-min-q" parameters from 1..62 to 0..63.

Assert that i is valid for the final "d[i++] = nullptr;" statement.