Change "constant QP" to "constant rate factor"
If rate_control_mode is set to 0 (the default), the rate control mode
depends on the value of enable_adaptive_quantization. If
enable_adaptive_quantization is 2 (the default), the rate control mode
is constant rate factor (CRF). If enable_adaptive_quantization is 0, the
rate control mode is constant QP. Since we don't set
enable_adaptive_quantization, the rate control mode is CRF.