uncompressed: split up uncompressed implementation
This moves the enumerations into a shared header file. The
sampling type enumeration and interleave type enumeration are
renamed to be sampling mode and enumeration mode to match the changes
in the FDIS. Documentation is added (although its not public API).

The cmpd and uncC box definitions are split off from the codec, to
keep file sizes reasonable.