fix wrong include directories for libde265 and aom in build-emscripten.sh
Currently when running the build-emscripten.sh script from another
directory (as the script suggests) the base path for libde265 set in
$LIBARY_LINKER_FLAGS and $LIBRARY_INCLUDE_FLAGS for get set to
(if the src dir of the project is / and the build dir is /buildjs)
/libde265-1.0.12 instead of /buildjs/libde265-1.0.12. This causes
the compiler to fail to find libde265.

This isn't caught in ./scripts/run-ci.sh because the script (on line 246)
has "source ./emscripten/emsdk/emsdk_env.sh && USE_WASM=0 ./build-emscripten.sh ."
using the same directory as both the source directory and the build directory.

make the same fixes for aom