Webcodecs: Handle limited range pixel values in NV12 data.
"Full range" means that the pixel data ranges from
0 to 255. "Limited range" means that it's using a
more constrained range, such as 16 to 235.

Because the range is being clamped, it prevents
images decoded with the VideoDecoder API from
using the full range of color.

Notably, this caused fully transparent backgrounds
to be light gray due to the bottom end of the
range starting at 16 instead of 0.

Also, explicilty handle the mono image channel
used to encode transparency information in HEIC
images.

Lastly, add the copyright header to
decoder_webcodecs.h.