Convert limited range NV12 YCbCr data to full range.
The browser VideoDecoder API may return limited
range color data even when the input source is
using full range.

The webcodecs plugin corrects for this by
inspecting whether the output is encoded in full
range or not.

Note that if the original source is using limited
range that this scaling could cause libheif to
scale the data again when converting to RGB. This
isn't easy to avoid at this time b/c plugins don't
know whether the input source is using limited
range or not. While it is technically present in
the VUI SPS data, it is rather complicated to
parse this structure as a one-off.